// types/type.ts
export type Attributes = {
  width: string;
  height: string;
  fontSize: string;
  fontFamily: string;
  fontWeight: string;
  fill: string;
  stroke: string;
  opacity: number;
  strokeWidth:string;
  strokeOpacity:number;
};

export type ActiveElement = {
  name: string;
  value: string;
  icon: string;
} | null;